Refactor server modules
