+++ packages/experiments/demo/server/dom.ts
@@ -1,0 +1,18 @@
+import * as jsdom from 'jsdom';
+import { html } from './html';
+
+export const dom = new jsdom.JSDOM(html);
+const document = dom.window.document;
+document.getSelection = () => ({}) as Selection;
+const globalAny:any = global;
+globalAny.document = document;
+globalAny.navigator = {};
+globalAny.window = {
+  addEventListener: () => {}
+};
+globalAny.MutationObserver = () => ({
+  observe: () => {},
+  takeRecords: () => {},
+  disconnect: () => {}
+});
+globalAny.requestAnimationFrame = () => {};
