+++ packages/experiments/demo/server/server.ts
@@ -1,14 +1,11 @@
 import * as express  from 'express';
-import { Application, Router, Request, Response } from 'express';
-import * as jsdom from 'jsdom';
 import * as http from 'http';
-import { createView } from '../demoView';
 
 import * as ShareDB from 'sharedb';
 import * as WebSocket from 'ws';
 import * as WebSocketJSONStream from '@teamwork/websocket-json-stream';
 
-import { html } from './html';
+import { router } from './router';
 
 const backend = new ShareDB();
 createDoc(startServer);
@@ -21,43 +18,15 @@ function createDoc(callback) {
       throw err;
     }
     if (doc.type === null) {
-      doc.create('', callback);
+      doc.create('Test content', callback);
       return;
     }
     callback();
   });
 }
 
-const { JSDOM } = jsdom;
-
-const dom = new JSDOM(html);
-const document = dom.window.document;
-document.getSelection = () => ({}) as Selection;
-const globalAny:any = global;
-globalAny.document = document;
-globalAny.navigator = {};
-globalAny.window = {
-  addEventListener: () => {}
-};
-globalAny.MutationObserver = () => ({
-  observe: () => {},
-  takeRecords: () => {},
-  disconnect: () => {}
-});
-globalAny.requestAnimationFrame = () => {};
-const view = createView({});
-document.querySelector("#editor").appendChild(view.dom);
-
 function startServer() {
-
-  const router: Router = Router();
-
-  router.get('/', (req: Request, res: Response) => {
-    const html = dom.serialize();
-    res.send(html);
-  });
-
-  const app: Application = express();
+  const app = express();
   app.use('/build', express.static('demo/build'));
   app.use('/', router);
